Revert "add push-to-light-thread dir, and upload mbed-os"
