+++ mbed_app.json
@@ -1,0 +1,7 @@
+{
+    "target_overrides": {
+        "*": {
+            "platform.thread-stats-enabled": true
+        }
+    }
+}
