--- push-to-light-thread/mbed-os.lib
@@ -1,1 +1,0 @@
-https://github.com/ARMmbed/mbed-os/
